add options from menu
